+++ src/packages/property-editors/toggle/manifests.ts
