+++ src/packages/property-editors/toggle/property-editor-ui-toggle.element.ts
@@ -1,4 +1,4 @@
-import type { UmbInputToggleElement } from '../../../components/input-toggle/input-toggle.element.js';
+import type { UmbInputToggleElement } from '../../core/components/input-toggle/input-toggle.element.js';
 import { html, customElement, property, state } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
 import { UmbPropertyValueChangeEvent } from '@umbraco-cms/backoffice/property-editor';
